# join-quiz/src/main/java/com/example/join_quiz/dto/config/ConfigOutputDto.java
package com.example.join_quiz.dto.config;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.experimental.Accessors;

import java.time.LocalTime;
import java.time.OffsetDateTime;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Accessors(chain = true)
public class ConfigOutputDto {

    private Long id;

    private OffsetDateTime startTime;

    private Long duration;
}