# quiz/src/main/java/com/example/quiz/dto/QuizOutputDto.java
package com.example.quiz.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.experimental.Accessors;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Accessors(chain = true)
public class QuizOutputDto {
    private Long id;
    private String name;
    private String description;
    private Long configId;
    private String authorName;
}